src/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.4
sentence-transformers>=2.7.0
torch>=2.2.0
numpy>=2.0.0
scikit-learn>=1.4.2
xgboost>=3.2.0
joblib>=1.3.2
python-dotenv>=1.0.1
requests>=2.31.0
jinja2>=3.1.2
aiofiles>=23.2.1
python-multipart>=0.0.9
